requirements.txt
python-dotenv>=1.0.0
mistralai>=0.4.0
flask>=3.0.0
streamlit>=1.30.0
requests>=2.31.0